requirement.txt
opencv-python
torchvision
pandas
tensorboard
torch
Pillow
easydict
matplotlib
scikit-learn
kornia
torchgeometry
lpips
tensorflow-io
tensorflow==2.5.0
setuptools==59.5.0

## also include the following two packages.
## git clone this one in your folder (https://github.com/mlomnitz/DiffJPEG)
## pip3 install git+https://github.com/fra31/auto-attack